better

+++ jpa/src/main/java/jdbc/job/InputDTO.java
@@ -12,10 +12,10 @@ public class InputDTO {
     private String firstName;
     private String lastName;
     private int age;
-    private long salary;
+    private int salary;
 
     private static final Faker faker = new Faker();
-    private static final AtomicInteger ids =new AtomicInteger(0);
+    private static final AtomicInteger ids = new AtomicInteger(0);
 
     public static InputDTO generate() {
         InputDTO inputDTO = new InputDTO();
